+++ packages/default-catalog/api/catalog.ts
@@ -5,6 +5,8 @@ import { adjustBackendProxyUrl } from '@storefront-api/lib/elastic'
 import cache from '@storefront-api/lib/cache-instance'
 import { sha3_224 } from 'js-sha3'
 import Logger from '@storefront-api/lib/logger'
+import bodybuilder from 'bodybuilder'
+import { elasticsearch, SearchQuery } from 'storefront-query-builder'
 
 function _cacheStorageHandler (config, result, hash, tags) {
   if (config.server.useOutputCache && cache) {
@@ -18,7 +20,23 @@ function _cacheStorageHandler (config, result, hash, tags) {
   }
 }
 
-export default ({config, db}) => function (req, res, body) {
+function _outputFormatter (responseBody, format = 'standard') {
+  if (format === 'compact') { // simple formatter
+    delete responseBody.took
+    delete responseBody.timed_out
+    delete responseBody._shards
+    if (responseBody.hits) {
+      delete responseBody.hits.max_score
+      responseBody.total = responseBody.hits.total
+      responseBody.hits = responseBody.hits.hits.map(hit => {
+        return Object.assign(hit._source, { _score: hit._score })
+      })
+    }
+  }
+  return responseBody
+}
+
+export default ({config, db}) => async function (req, res, body) {
   let groupId = null
 
   // Request method handling: exit if not GET or POST
@@ -27,14 +45,18 @@ export default ({config, db}) => function (req, res, body) {
     throw new Error('ERROR: ' + req.method + ' request method is not supported.')
   }
 
-  let requestBody: Record<string, any> = {}
+  let requestBody: Record<string, any> = req.body as Record<string, any>
+  let responseFormat = 'standard'
   if (req.method === 'GET') {
     if (req.query.request) { // this is in fact optional
       requestBody = JSON.parse(decodeURIComponent(req.query.request))
     }
-  } else {
-    requestBody = req.body
   }
+
+  if (req.query.request_format === 'search-query') { // search query and not Elastic DSL - we need to translate it
+    requestBody = await elasticsearch.buildQueryBodyFromSearchQuery({ config, queryChain: bodybuilder(), searchQuery: new SearchQuery(requestBody) })
+  }
+  if (req.query.response_format) responseFormat = req.query.response_format
 
   const urlSegments = req.url.split('/');
 
@@ -108,6 +130,7 @@ export default ({config, db}) => function (req, res, body) {
         if (entityType === 'product') {
           resultProcessor.process(_resBody.hits.hits, groupId).then((result) => {
             _resBody.hits.hits = result
+            _resBody = _outputFormatter(_resBody, responseFormat)
             _cacheStorageHandler(config, _resBody, reqHash, tagsArray)
             res.json(_resBody);
           }).catch((err) => {
@@ -116,6 +139,7 @@ export default ({config, db}) => function (req, res, body) {
         } else {
           resultProcessor.process(_resBody.hits.hits).then((result) => {
             _resBody.hits.hits = result
+            _resBody = _outputFormatter(_resBody, responseFormat)
             _cacheStorageHandler(config, _resBody, reqHash, tagsArray)
             res.json(_resBody);
           }).catch((err) => {