+++ packages/default-catalog/graphql/elasticsearch/attribute/resolver.ts
@@ -4,14 +4,14 @@ import { buildQuery } from '../queryBuilder';
 import { getIndexName } from '../mapping'
 import { adjustQuery, getResponseObject } from '@storefront-api/lib/elastic'
 
-async function listAttributes ({ attributes = null, filter = null, context, rootValue, _sourceInclude, _sourceExclude }) {
+async function listAttributes ({ attributes = null, filter = null, context, rootValue, _sourceIncludes, _sourceExcludes }) {
   let query = buildQuery({ filter: filter || attributes, pageSize: 150, type: 'attribute' });
 
   const esQuery = {
     index: getIndexName(context.req.url),
     body: query,
-    _source_include: _sourceInclude,
-    _source_exclude: _sourceExclude
+    _source_includes: _sourceIncludes,
+    _source_excludes: _sourceExcludes
   }
 
   const response = getResponseObject(await client.search(adjustQuery(esQuery, 'attribute', config)));
@@ -26,11 +26,11 @@ async function listAttributes ({ attributes = null, filter = null, context, root
   return response;
 }
 
-export async function listSingleAttribute ({ attribute_id, attribute_code, context, rootValue, _sourceInclude, _sourceExclude }) {
+export async function listSingleAttribute ({ attribute_id, attribute_code, context, rootValue, _sourceIncludes, _sourceExcludes }) {
   const filter = {}
   if (attribute_id) filter['attribute_id'] = { eq: attribute_id }
   if (attribute_code) filter['attribute_code'] = { eq: attribute_code }
-  const attrList = await listAttributes({ filter, context, rootValue, _sourceInclude, _sourceExclude })
+  const attrList = await listAttributes({ filter, context, rootValue, _sourceIncludes, _sourceExcludes })
   if (attrList && attrList.items.length > 0) {
     return attrList.items[0]
   } else {
@@ -40,8 +40,8 @@ export async function listSingleAttribute ({ attribute_id, attribute_code, conte
 
 const resolver = {
   Query: {
-    customAttributeMetadata: (_, { attributes, _sourceInclude, _sourceExclude }, context, rootValue) => listAttributes({ attributes, context, rootValue, _sourceInclude, _sourceExclude }),
-    attribute: (_, { attribute_id, attribute_code, _sourceInclude, _sourceExclude }, context, rootValue) => listSingleAttribute({ attribute_id, attribute_code, _sourceInclude, _sourceExclude, context, rootValue })
+    customAttributeMetadata: (_, { attributes, _sourceIncludes, _sourceExcludes }, context, rootValue) => listAttributes({ attributes, context, rootValue, _sourceIncludes, _sourceExcludes }),
+    attribute: (_, { attribute_id, attribute_code, _sourceIncludes, _sourceExcludes }, context, rootValue) => listSingleAttribute({ attribute_id, attribute_code, _sourceIncludes, _sourceExcludes, context, rootValue })
   }
 };