+++ src/index.ts
@@ -5,7 +5,7 @@ import app, { hasGraphqlSupport } from './server';
 const port = process.env.PORT || config.get<string>('server.port')
 const host = process.env.HOST || config.get<string>('server.host')
 app.listen(parseInt(port), host, () => {
-  Logger.info(`Storefront GraphQL API started at http://${host}:${port}`);
+  Logger.info(`Storefront REST API started at http://${host}:${port}`);
   if (hasGraphqlSupport) {
     Logger.info(`Storefront GraphQL API started at http://${host}:${port}/graphql`);
   }